Conveyor
Moved conveyor out of constants to its own file
