+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -14,9 +14,7 @@ import edu.wpi.first.wpilibj.PneumaticsModuleType;
 import edu.wpi.first.wpilibj.Solenoid;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
-import frc.robot.Constants.DrivetrainConstants.Conveyor;
-import frc.robot.Constants.DrivetrainConstants.IntakeConstants;
+import frc.robot.Constants.IntakeConstants;
 
 /** Add your docs here. */
